Create state dir and move game_state.rb
