+++ game_gui.rb
@@ -5,7 +5,7 @@ require './gui_elements/dialog.rb'
 require './game.rb'
 require './gui_input_manager.rb'
 require './game_driver.rb'
-require './game_state.rb'
+require './state/game_state.rb'
 
 class GameGui < Gosu::Window
     def initialize(logger, prompt_strings, user_prompt_templates, deck)
