+++ state/game_state.rb
